# Ignore Optimizer.step nodes covering >90% runtime, plan to generalize later